Properly calculate shares/assets in withdraw and redeem (#4)
**Changes in this PR**
- Improve calculation of shares and assets
  - Calculate the shares/assets locally instead of making two contract calls to `Comet`
  - Local calculations allow us to delay transferring out assets until the end of the function to avoid potential re-entrancy attacks
  - `previewRedeem/Withdraw` now return exact amounts, where the previous implementation would give values that could be off by 2 gwei
  - Avoid having to hardcode any logic (e.g. "subtract 1 from shares to redeem") by always rounding up or down in the wrapper's favor
- Make `redeem` conform to the ERC-4626 spec by making it burn exact `shares` instead of `shares-1`
- Remove all uses of named return variables
- Add invariant tests for `redeem` to ensure correct accounting of shares

**Explanation**
The old logic for calculating 1) `assets` to withdraw in `redeem` and  2) `shares` to redeem in `withdraw` is pretty wonky. 

It makes two contract calls to Comet to get the `principal` value pre and post transfer in order to calculate the diff. This diff in `principal` can actually be calculated locally, which helps to save a good amount of gas and allows us to transfer assets out at the end of the function to avoid any potential re-entrancy attacks. The old logic for `redeem` also does not conform to the ERC-4626 spec, as the function does not burn the exact amount of `shares` specified by the caller (it burns `shares - 1` instead). This lead to weird rounding behaviors, where some balances were off by 2 wei instead of 1 wei.

The new logic for shares/assets calculation improves upon the old implementation by calculating all the values locally to gain all the benefits discussed above. It also makes sure that `redeem` now conforms to the ERC-4626 spec and will always burn the exact amount of `shares` specified by the caller.